SMV: define FILE_SIZE macro to define unsigned 4 byte ints on 32 bit systems and unsigned 8 byte longs on 64 bit systems .  Used to pass character strings from C to Fortran
Former-commit-id: 62e1d2f20208f916e33d2e839325cdae287506e7 [formerly 38bf2f12ec9fd86fe0818cacecc2a419029bea46]
Former-commit-id: 37f30fc450063386d50a0aba17b7fe97c3df3ca5
Former-commit-id: a87d6fb515d57506dcdb7374f77caddae585f05c

+++ SMV_5/source/smokeview/dmalloc.c
@@ -31,7 +31,7 @@ void _memorystatus(unsigned int size,unsigned int *availmem,unsigned int *physme
 #ifdef _DEBUG
     if(size!=0&&size<=stat.dwAvailPhys-0.1*stat.dwTotalPhys){
       printf("*** Available Memory: %i M \n",
-            stat.dwAvailPhys/(1024*1024));
+           (int)stat.dwAvailPhys/(1024*1024));
     }
 #endif
     if(size!=0&&size>stat.dwAvailPhys-0.1*stat.dwTotalPhys){
